# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "UXUS — Agent Services",
    "version": "1.2.0",
    "description": "Pay-per-call primitives for AI agents. LLM inference, web scraping, structured extraction, embeddings, live search, and EVM wallet/token risk scoring. Paid in USDC on Base via the x402 protocol (v2). No accounts, no API keys.",
    "contact": {
      "name": "UXUS",
      "url": "https://uxus.finance"
    }
  },
  "servers": [
    {
      "url": "https://uxus.finance"
    }
  ],
  "components": {
    "securitySchemes": {
      "x402": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "x402-payment",
        "description": "x402 payment, protocol v2. An unpaid request returns 402 with the payment requirement base64-encoded in the PAYMENT-REQUIRED response header and mirrored as JSON in the body. Network: Base (eip155:8453). Asset: USDC (0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913). Pay to 0xe0ed7a30589fec49e98f2085c7162b90fdbb83de. Sign with an x402-capable client and retry with the X-PAYMENT header."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "bad_request"
          },
          "detail": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Launch": {
        "type": "object",
        "properties": {
          "pair": {
            "type": "string",
            "nullable": true
          },
          "tokenAddress": {
            "type": "string",
            "nullable": true
          },
          "priceUsd": {
            "type": "number",
            "nullable": true
          },
          "liquidityUsd": {
            "type": "number"
          },
          "fdvUsd": {
            "type": "number",
            "nullable": true
          },
          "volumeH24Usd": {
            "type": "number",
            "nullable": true
          },
          "ageMinutes": {
            "type": "integer"
          },
          "flags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dex": {
            "type": "string",
            "nullable": true
          },
          "riskCheck": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "LaunchesResponse": {
        "type": "object",
        "required": [
          "chain",
          "count",
          "launches"
        ],
        "properties": {
          "chain": {
            "type": "string",
            "enum": [
              "ethereum",
              "base"
            ]
          },
          "count": {
            "type": "integer"
          },
          "launches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Launch"
            }
          },
          "note": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "example": "geckoterminal"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LookupResult": {
        "type": "object",
        "required": [
          "address",
          "chain",
          "chain_detected",
          "exists",
          "risk_score",
          "risk_level",
          "verdict",
          "flags",
          "entity",
          "name"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "chain": {
            "type": "string",
            "enum": [
              "evm",
              "xrpl"
            ]
          },
          "chain_detected": {
            "type": "boolean"
          },
          "exists": {
            "type": "boolean",
            "nullable": true,
            "description": "null when on-chain state could not be verified (EVM RPC failure, or XRPL node unreachable). Never guessed."
          },
          "risk_score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "nullable": true,
            "description": "null (XRPL) when the account state could not be read — not scored, not clean"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "nullable": true
          },
          "verdict": {
            "type": "string",
            "enum": [
              "PROCEED",
              "CAUTION",
              "BLOCK"
            ],
            "nullable": true
          },
          "flags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "entity": {
            "$ref": "#/components/schemas/Entity"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "description": "primary ENS/Basename for an EVM address, forward-verified; null if none"
          },
          "degraded": {
            "type": "boolean",
            "description": "true when a verdict input was unavailable: XRPL node unreachable (exists/risk come back null), or EVM wallet-history lookup failed (verdict is CAUTION not PROCEED, with a HISTORY_UNAVAILABLE flag). Retry before acting."
          },
          "note": {
            "type": "string",
            "description": "present when degraded (XRPL): what could not be checked"
          },
          "upgrade": {
            "type": "string",
            "nullable": true,
            "description": "points at /api/risk/pro for EVM; null for XRPL"
          },
          "disclaimer": {
            "type": "string",
            "description": "Ships inside the response: developer-grade screening from public data, not legal/financial/compliance advice. See lib/disclaimer.ts."
          }
        }
      },
...
      "Entity": {
        "type": "object",
        "required": [
          "is_known",
          "label",
          "category"
        ],
        "properties": {
          "is_known": {
            "type": "boolean"
          },
          "label": {
            "type": "string",
            "nullable": true,
            "description": "human name, e.g. \"Tornado.Cash: 1 ETH\"; null when unknown"
          },
          "category": {
            "type": "string",
            "description": "\"sanctioned\" is set only when the address is on the OFAC SDN list (the single source of truth, shared with /api/risk). \"flagged\" means a non-OFAC blocklist named it — informational, not a hard stop.",
            "enum": [
              "sanctioned",
              "mixer",
              "exchange",
              "bridge",
              "dex_router",
              "scam",
              "drainer",
              "phishing",
              "flagged",
              "token_contract",
              "protocol",
              "unknown"
            ]
          },
          "sanctions_source": {
            "type": "string",
            "description": "present only when category is \"sanctioned\": the list that designated it (currently always \"ofac\")"
          }
        }
      },
      "EntityResult": {
        "type": "object",
        "required": [
          "address",
          "chain",
          "is_known",
          "label",
          "category"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "chain": {
            "type": "string",
            "enum": [
              "ethereum",
              "base"
            ]
          },
          "is_known": {
            "type": "boolean"
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "enum": [
              "sanctioned",
              "mixer",
              "exchange",
              "bridge",
              "dex_router",
              "scam",
              "drainer",
              "phishing",
              "flagged",
              "token_contract",
              "protocol",
              "unknown"
            ]
          },
          "sanctions_source": {
            "type": "string",
            "description": "present only when category is \"sanctioned\": the list that designated it (currently always \"ofac\", from the OFAC SDN list in bad_addresses)"
          },
          "disclaimer": {
            "type": "string",
            "description": "Ships inside the response: developer-grade screening from public data, not legal/financial/compliance advice. See lib/disclaimer.ts."
          }
        }
      },
      "CallerIdResult": {
        "type": "object",
        "required": [
          "address",
          "chain",
          "name",
          "entity",
          "exists",
          "risk_score",
          "risk_level",
          "verdict",
          "recommendation",
          "confidence",
          "reasons",
          "checked_at"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "chain": {
            "type": "string",
            "enum": [
              "ethereum",
              "base"
            ]
          },
          "name": {
            "type": "string",
            "nullable": true,
            "description": "forward-verified primary ENS/Basename, or null"
          },
          "entity": {
            "$ref": "#/components/schemas/Entity"
          },
          "exists": {
            "type": "boolean",
            "nullable": true,
            "description": "null when on-chain state could not be checked"
          },
          "risk_score": {
            "type": "integer",
            "nullable": true,
            "minimum": 0,
            "maximum": 100
          },
          "risk_level": {
            "type": "string",
            "nullable": true,
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ]
          },
          "verdict": {
            "type": "string",
            "nullable": true,
            "enum": [
              "PROCEED",
              "CAUTION",
              "BLOCK"
            ]
          },
          "risk_degraded": {
            "type": "boolean",
            "nullable": true,
            "description": "true = risk was scored but wallet history was unavailable; verdict is CAUTION not PROCEED, and confidence is 'low'"
          },
          "recommendation": {
            "type": "string",
            "enum": [
              "ANSWER",
              "SCREEN",
              "BLOCK"
            ]
          },
          "confidence": {
            "type": "string",
            "enum": [
              "high",
              "low"
            ],
            "description": "high = entity and a fully-assessed risk both resolved; low = an input was missing OR the risk read was degraded (treat as unchecked, not clean)"
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "short strings naming exactly what drove the recommendation"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time"
          },
          "disclaimer": {
            "type": "string",
            "description": "Ships inside the response: developer-grade screening from public data, not legal/financial/compliance advice. See lib/disclaimer.ts."
          }
        }
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Invalid or missing parameters.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "UpstreamError": {
        "description": "An upstream provider failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/lookup": {
      "get": {
        "operationId": "universalLookup",
        "summary": "FREE universal reverse lookup — any address, any chain",
        "description": "Auto-detects the chain from the address format (EVM 0x address or XRPL classic r-address) — no chain parameter. One normalized envelope for every chain. EVM screening defaults to Base; ?chain=ethereum moves the wallet-history signal only (sanctions/scam lists are chain-agnostic). XRPL X-addresses are recognized but return 400 asking for the classic form.",
        "security": [],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045"
          },
          {
            "name": "chain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "base",
                "ethereum"
              ],
              "default": "base"
            },
            "description": "Optional EVM history-signal hint. Ignored for XRPL."
          }
        ],
        "responses": {
          "200": {
            "description": "Normalized lookup result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/callerid": {
      "get": {
        "operationId": "callerId",
        "summary": "FREE wallet caller ID — should this wallet be answered",
        "description": "For wallet messaging (XMTP, Push), which has no spam filter. Composes entity attribution + risk score + on-chain history + forward-verified name into ANSWER | SCREEN | BLOCK with reasons[]. BLOCK on entity category sanctioned/drainer/phishing/scam/mixer or risk verdict BLOCK. ANSWER on a known exchange/dex_router/protocol OR a forward-verified ENS/Basename, each with risk verdict PROCEED. SCREEN is the default for unknown addresses. confidence is 'low' when entity or risk could not be checked — degrades to SCREEN, never guesses.",
        "security": [],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "0x47CE0C6eD5B0Ce3d3A51fdb1C52DC66a7c3c2936"
          },
          {
            "name": "chain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "base",
                "ethereum"
              ],
              "default": "base"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Caller-ID recommendation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallerIdResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/api/entity": {
      "get": {
        "operationId": "entityAttribution",
        "summary": "FREE entity attribution — what IS this address",
        "description": "Answers what an address is (not who owns it) from ScamSniffer, MEW, eth-labels and curated Base/Tornado sets. The \"sanctioned\" category comes solely from the OFAC SDN list ingested into bad_addresses — the same source /api/risk uses; a non-OFAC blocklist hit is category \"flagged\", not \"sanctioned\". is_known:false / category:\"unknown\" when nothing matches — never inferred from address shape.",
        "security": [],
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "0x47CE0C6eD5B0Ce3d3A51fdb1C52DC66a7c3c2936"
          },
          {
            "name": "chain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "base",
                "ethereum"
              ],
              "default": "base"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entity attribution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/api/launches": {
      "get": {
        "operationId": "launches",
        "summary": "FREE fresh DEX pool feed",
        "description": "Newest DEX pools on Base/Ethereum (GeckoTerminal). Each row carries a riskCheck link into /api/risk/live.",
        "security": [],
        "parameters": [
          {
            "name": "chain",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ethereum",
                "base"
              ],
              "default": "base"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 30,
              "default": 20
            }
          },
          {
            "name": "minLiquidityUsd",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fresh pools",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaunchesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "502": {
            "$ref": "#/components/responses/UpstreamError"
          }
        }
      }
    }
  }
}